% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accuracy_metrics}
\alias{accuracy_metrics}
\title{Calcul des Métriques combinées}
\usage{
accuracy_metrics(forecast_val, real_val)
}
\arguments{
\item{forecast_val}{Vecteur des valeurs prédites.}

\item{real_val}{Vecteur des valeurs réelles.}
}
\value{
Un vecteur nommé avec RMSE et MAPE.
}
\description{
Calcul des Métriques combinées
}
