% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_mape}
\alias{calculate_mape}
\title{Calcul du MAPE (Mean Absolute Percentage Error)}
\usage{
calculate_mape(actual, predicted)
}
\arguments{
\item{actual}{Valeurs réelles}

\item{predicted}{Valeurs prédites}
}
\description{
Calcul du MAPE (Mean Absolute Percentage Error)
}
