% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_rmse}
\alias{calculate_rmse}
\title{Calcul du RMSE (Root Mean Squared Error)}
\usage{
calculate_rmse(actual, predicted)
}
\arguments{
\item{actual}{Valeurs réelles}

\item{predicted}{Valeurs prédites}
}
\description{
Calcul du RMSE (Root Mean Squared Error)
}
