% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_elec_data}
\alias{clean_elec_data}
\title{Nettoyage des données (Logique Partie 1)}
\usage{
clean_elec_data(numeric_vec)
}
\arguments{
\item{numeric_vec}{Un vecteur numérique (la série temporelle).}
}
\value{
Le vecteur nettoyé.
}
\description{
Remplace les valeurs inférieures à 130 kW par la moyenne des valeurs
aux mêmes heures les 3 jours précédents.
}
